# spec/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/batches/{batchId}/images:
    put:
      description: Append uploaded images to the batch (max 5 total). Rejects if the batch already has 5 images.
      operationId: BatchIdImagesController_addImages_v1
      parameters:
        - name: batchId
          required: true
          in: path
          description: Batch ID
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Image files (JPEG or PNG, max 5MB each)
                  maxItems: 5
              required:
                - images
      responses:
        '200':
          description: Images added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          description: Invalid file type, file too large, or invalid filename
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Batch or image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '409':
          description: Batch already has the maximum of 5 images
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '415':
          description: Invalid content type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Upload batch images
      tags:
        - batches
        - images
  /v1/batches/{batchId}/images/{filename}:
    delete:
      operationId: BatchIdImagesController_removeImage_v1
      parameters:
        - name: filename
          required: true
          in: path
          schema:
            type: string
        - name: batchId
          required: true
          in: path
          description: Batch ID
          schema:
            type: integer
            minimum: 1
            example: 42
      responses:
        '204':
          description: Image removed successfully
        '400':
          description: Invalid file type, file too large, or invalid filename
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Batch or image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '406':
          description: Database-related error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '409':
          description: Batch already has the maximum of 5 images
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '415':
          description: Invalid content type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 422
                  error:
                    type: string
                    example: Unprocessable Entity
                  message:
                    type: string
                    example:
                      - propertyA has wrong value "...", [constrains]
                      - propertyB has wrong value "...", [constrains]
                      - '...'
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '500':
          description: Generic unknown error produced by an unidentified exception
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 500
                  error:
                    type: string
                    example: Internal Server Error
                  message:
                    type: string
                    example: Internal server error
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
        '503':
          description: The event loop is lagging and no more requests can be attended right now
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: number
                    example: 503
                  error:
                    type: string
                    example: Service Unavailable
                  message:
                    type: string
                    example: Server is too busy
                    description: The sort of error message you can expect
                  emitter:
                    type: string
                    example: ErrorProducerModule
                    description: If the error was produced within the codebase, the name of the module where it was created
                  timestamp:
                    type: string
                    example: '2024-09-14T22:00:00.000Z'
                  path:
                    type: string
                    example: /path/of/the/failing/route
                    description: Endpoint responsible for the production of the error
      summary: Remove a batch image by filename
      tags:
        - batches
        - images
...
info:
  title: mushpi-server
  description: A backend for the Mushroom Pi project
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    secret:
      type: http
      scheme: bearer
      in: header
      name: authorization
  schemas:
...
    ErrorDto:
      type: object
      properties: {}
...
    Batch:
      type: object
      properties:
        images:
          description: Filenames of attached images
          type: array
          items:
            type: string
        images_url:
          description: Absolute URLs to the batch images (computed)
          example:
            - http://localhost:3000/images/batches/7/1.jpg
          type: array
          items:
            type: string
        status:
          type: string
          enum:
            - planned
            - in-progress
            - finished
          description: Current status of the batch (computed from start_at and finish_at)
          example: in-progress
        images_left:
          type: number
          description: Number of images that can still be uploaded to this batch
          example: 3
          minimum: 0
          maximum: 5
      required:
        - status
        - images_left
...